fix: Only jump each axis if it's enabled
This fixes the bug where the upper right corner of the screen will jump
to the bottom left corner when only x is enabled.
Also fixes the bug where the lower left corner of the screen will jump
to the top right corner when only y is enabled.

This also changes the `axis` variable to be a bit mask based on the new
`axis_t` enum. This makes it easier to check if an axis is enabled,
instead of having to check for both `X_AXIS` and `BOTH_AXES`.